refactor: Avoid re-renders - UnsavedChangesModal moved to its own component for avoiding re-render whole page when modal is opening
